Yes.
**Updated the UI**
Added Search Bar!
Added Favorites Section!
On the Bottom right corner of the footer area, there will now display the **YOUR** Current API Version, There will be a orange dot next to it, IF there is a new API Version, If that happens, Please Refresh your page!
Fixed the Typos...